+++ client/app.jsx
@@ -1,15 +1,15 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch} from "react-router-dom";
 import NavBar from "./components/Navbar.jsx"
-import DemoContainer from "./containers/DemoContainer.jsx"
+import Demo from "./components/Demo.jsx"
 
 
 const App = () => {
   return (
     <Router>
-      <NavBar />
+      {/* <NavBar /> */}
       <Switch>
-        <DemoContainer />
+        <Demo />
       </Switch>
     </Router>
   );
