+++ client/index.html
@@ -6,10 +6,6 @@
     <title>Document</title>
   </head>
   <body>
-    <button class="queries" id="query1">Get Human Query (POST)</button>
-    <button class="queries" id="query2">Get Human Query (GET)</button>
-    <button class="queries" id="query3">Simple Pokemon Query (POST)</button>
-    <button class="queries" id="query4"> Nested Pokemon Query (POST)</button>
-    <button class="mutations" id="mutation1">Add Human Mutation (POST)</button>
+    <div id="root"></div>
   </body>
 </html>
