+++ client/index.js
@@ -3,6 +3,8 @@ import React from 'react';
 import { render } from 'react-dom';
 import App from './app.jsx';
 
-// User passes in configuration object into registration function
-// The function will save these settings to IDB so they can be accessed by the service worker
-register({ useMetrics: true, cacheExpirationLimit: 1000, doNotCache: ['capture_rate']});
+render(
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+  , document.getElementById('root'));
