+++ package.json
@@ -28,7 +28,6 @@
     "graphql": "^15.5.1",
     "graphql-tag": "^2.12.5",
     "html-webpack-plugin": "^5.3.2",
-    "idb-keyval": "^5.1.3",
     "lowdb": "^1.0.0",
     "nanoid": "^2.1.1",
     "path": "^0.12.7",
