+++ webpack.config.js
@@ -7,6 +7,7 @@ const __dirname = dirname(fileURLToPath(import.meta.url)); // https://stackoverf
 export default {
   entry: {
     index: './client/index.js',
+    handlers: './client/handlers.js',
     sw: './node_modules/loql/sw.js',
   },
   devtool: 'eval-source-map',
@@ -23,8 +24,8 @@ export default {
         exclude: /node_modules/,
         loader: 'babel-loader',
         options: {
-          presets: ['@babel/preset-react'],
-          // plugins: ['@babel/plugin-transform-runtime', '@babel/transform-async-to-generator'],
+          presets: ['@babel/preset-env', '@babel/preset-react'],
+          plugins: ['@babel/plugin-transform-runtime', '@babel/transform-async-to-generator'],
         },
       },
       {
